Implementing equals in StudentCurriculumModuleBean
It's interesting to implement equals (and hashCode) in these beans so
HashMaps, Sets and Caches (among others) may take advantage of them.

My idea was to consider that if two Beans are constructed exactly with
the same parameters they are the same. Not sure if there's some side
behaviour that may render this premise false.