Changing the use of getExternalId for getOid
Although externalId should always be the one used for the sake of
portability, if we compare things via externalId to ensure we can always
sort domain objects we will be converting from Long to String and then
perform string comparation. If we use directly the Oid it will be a
simple Long comparation.

Specially now in ff 3.0 that getOid() is a Comparable and not a Long
it's, in my opinion, even less problematic to use it in this sort of
scenario.

Relates with: #qubIT-Fenix-5453